fix: use BigDecimal for valorTotal to avoid floating-point precision issues
Agent-Logs-Url: https://github.com/Basilios-SPTech/Basilios---Projeto-de-PI-/sessions/757df9b8-b610-467b-af06-864fada571bd

+++ produtor/src/main/java/com/basilios/produtor/dto/PedidoDTO.java
@@ -4,6 +4,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record PedidoDTO(
@@ -14,7 +15,7 @@ public record PedidoDTO(
         List<@NotBlank String> itens,
 
         @Positive(message = "O valor total deve ser positivo")
-        double valorTotal,
+        BigDecimal valorTotal,
 
         String observacao
 ) {}
